+++ src/Symfony/Component/Validator/Tests/Constraints/CronTest.php
@@ -1,0 +1,57 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Validator\Tests\Constraints;
+
+use PHPUnit\Framework\TestCase;
+use Symfony\Component\Validator\Constraints\Cron;
+use Symfony\Component\Validator\Exception\ConstraintDefinitionException;
+use Symfony\Component\Validator\Mapping\ClassMetadata;
+use Symfony\Component\Validator\Mapping\Loader\AttributeLoader;
+
+class CronTest extends TestCase
+{
+    public function testAttributes()
+    {
+        $metadata = new ClassMetadata(CronDummy::class);
+        $loader = new AttributeLoader();
+        self::assertTrue($loader->loadClassMetadata($metadata));
+
+        [$bConstraint] = $metadata->getPropertyMetadata('b')[0]->getConstraints();
+        self::assertSame('myMessage', $bConstraint->message);
+        self::assertSame(['Default', 'CronDummy'], $bConstraint->groups);
+        self::assertSame(Cron::MODE_HASHED, $bConstraint->mode);
+
+        [$cConstraint] = $metadata->getPropertyMetadata('c')[0]->getConstraints();
+        self::assertSame(['my_group'], $cConstraint->groups);
+        self::assertSame('some attached data', $cConstraint->payload);
+    }
+
+    public function testUnknownModeThrows()
+    {
+        $this->expectException(ConstraintDefinitionException::class);
+        $this->expectExceptionMessage('The "invalid" validation mode is not supported.');
+
+        new Cron(mode: 'invalid');
+    }
+}
+
+class CronDummy
+{
+    #[Cron]
+    private $a;
+
+    #[Cron(message: 'myMessage', mode: Cron::MODE_HASHED)]
+    private $b;
+
+    #[Cron(groups: ['my_group'], payload: 'some attached data')]
+    private $c;
+}